+++ src/settings.ts
@@ -3,6 +3,7 @@ import * as vscode from 'vscode';
 class ExtensionSettings {
   private static readonly CONFIG_KEY: string = 'openRelatedFiles';
   private static readonly CONFIG_EXTENSIONS_KEY: string = 'limitToExtensions';
+  private static readonly CONFIG_OPEN_SINGLE_FILE: string = 'openSingleFile';
 
   private configuration: any;
 
@@ -20,6 +21,10 @@ class ExtensionSettings {
 
   shouldShowExtension(extension: string): boolean {
     return this.hasLimitToExtensions && this.limitToExtensions.indexOf(extension) > -1;
+  }
+
+  get openSingleFile(): boolean {
+    return this.configuration.get(ExtensionSettings.CONFIG_OPEN_SINGLE_FILE);
   }
 }