+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to the "open-related-file" extension will be documented in t
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [Unreleased] - 2018-09-15
+- Add `openSingleFile` option that defaults to true.
+- If there is only one related file, open it without prompting (if `openSingleFile` is true).
+
 ## [0.0.3] - 2018-09-14
 - Add better metadata